changes to make the recipient control component and address input component more readable
