+++ helpers/index.ts
@@ -1,4 +1,4 @@
-export { default as classNames } from './classNames';
+export * from './classNames';
 export * from './string';
 export * from './env';
 export * from './appVersion';
